pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sqlforensic"
version = "1.0.0"
description = "Database forensics toolkit — reverse-engineer undocumented databases in minutes"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Mehmet Can Diri"},
]
keywords = ["database", "forensics", "sql", "analysis", "schema", "documentation", "stored-procedures"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Software Development :: Documentation",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyodbc>=5.0",
    "psycopg2-binary>=2.9",
    "jinja2>=3.1",
    "sqlparse>=0.5",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "ruff>=0.2",
]

[project.scripts]
sqlforensic = "sqlforensic.cli:main"

[project.urls]
Homepage = "https://github.com/mehmetcandiri/sqlforensic"
Repository = "https://github.com/mehmetcandiri/sqlforensic"
Issues = "https://github.com/mehmetcandiri/sqlforensic/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
